First build of 2018!
If you look at the commit logs, a fair amount has changed in the month since my previous build.

This one has a hell of a lot of bug fixes, and is the version which has been used to build *all* of PI's internal and external services as of today.