Substantial build - use with care!
This build is fairly substantial, and will break a lot of things if you're not careful!  You should only use this version on a fresh build, unless you know what you're doing.

/media/VMs layout has changed to try and make it a little more obvious the difference between "data" and "storage".

This build also moves from VDIs as backing storage to VMDKs.

This is the first step towards dropping nbd for libguestfs for backups.  The backup process is very much WIP.